Fix import from init
