+++ cartoframes/data/__init__.py
@@ -3,14 +3,8 @@ from __future__ import absolute_import
 from .dataset.dataset import Dataset
 from .dataset.registry.strategies_registry import StrategiesRegistry
 from .clients.sql_client import SQLClient
-from .services import __all__ as services
-from .observatory import __all__ as observatory
-from .enrichment import __all__ as enrichment
 
 __all__ = [
-    'services',
-    'observatory',
-    'enrichment',
     'Dataset',
     'StrategiesRegistry',
     'SQLClient'
